monarch package: Adds build_runner dependency, fixes #161
When Monarch uses build_runner 2.5.x to build the preview-bundle
it fails as detailed in the issue.

This fix sets bounds around
the build_runner dependency by adding them to the monarch package.
This way, when `monarch init` adds the `build_runner` dependency,
it will be constrained by the version range declared in the
monarch package.